+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Bump firebase remote config to v21.1.2
 - Bump Google play services auth to v20.3.0
+- Bump datadog gradle plugin to v1.5.1
 
 ### Fixes