+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Bump firebase remote config to v21.1.2
 - Bump Google play services auth to v20.3.0
 - Bump datadog gradle plugin to v1.5.1
+- Bump Open CSV to v5.7.0
 
 ### Fixes